+++ rgbkbd/rgb_keyboard_controller_impl.h
@@ -80,7 +80,7 @@ class RgbKeyboardControllerImpl : public RgbKeyboardController {
 
   uint32_t GetRgbKeyboardCapabilities() override;
   void SetCapsLockState(bool enabled) override;
-  void SetStaticBackgroundColor(uint32_t r, uint32_t g, uint32_t b) override;
+  void SetStaticBackgroundColor(uint8_t r, uint8_t g, uint8_t b) override;
   void SetRainbowMode() override;
   void SetAnimationMode(RgbAnimationMode mode) override;
   void SetKeyboardClient(RgbKeyboard* keyboard) override;
